requirements.txt
adrf==0.1.6
asgiref==3.8.1
async-property==0.2.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-oauth-toolkit==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.2.0
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.8.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
priority==2.0.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.30.1
wsproto==1.2.0
